# openapi.yaml
openapi: 3.0.3
info:
  title: Joinly API
  version: 0.1.0
  description: API REST del MVP de eventos locales. Los recursos de producto requieren un JWT Bearer de Supabase Auth.
servers:
  - url: http://localhost:8080/api/v1
    description: Desarrollo local
security:
  - bearerAuth: []
paths:
...
  /events/{eventId}:
    parameters: [{$ref: '#/components/parameters/EventId'}]
    get:
      summary: Obtiene un evento visible
      responses:
        '200': {description: Evento, headers: {ETag: {$ref: '#/components/headers/ETag'}}, content: {application/json: {schema: {$ref: '#/components/schemas/EventDetail'}}}}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      summary: Edita un evento propio
      parameters: [{$ref: '#/components/parameters/IfMatch'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/EventPatch'}}}
      responses:
        '200': {description: Evento actualizado, headers: {ETag: {$ref: '#/components/headers/ETag'}}, content: {application/json: {schema: {$ref: '#/components/schemas/EventDetail'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
        '428': {$ref: '#/components/responses/PreconditionRequired'}
...
  /events/{eventId}/participations:
    parameters: [{$ref: '#/components/parameters/EventId'}]
    get:
      summary: Lista participantes de un evento propio (confirmados o pendientes)
      parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}, {name: status, in: query, schema: {type: string, enum: [confirmed, pending], default: confirmed}}]
      responses:
        '200': {description: Participantes, content: {application/json: {schema: {$ref: '#/components/schemas/ParticipantPage'}}}}
        '404': {$ref: '#/components/responses/NotFound'}
    post:
      summary: Crea una participación o solicitud
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        content: {application/json: {schema: {type: object, properties: {invitationCode: {type: string, minLength: 1, maxLength: 128}}}}}
      responses:
        '201': {description: Participación creada, headers: {ETag: {$ref: '#/components/headers/ETag'}}, content: {application/json: {schema: {$ref: '#/components/schemas/Participation'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /events/{eventId}/participation:
    parameters: [{$ref: '#/components/parameters/EventId'}]
    delete:
      summary: Abandona una participación confirmada
      responses:
        '204': {description: Participación abandonada}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /events/{eventId}/participations/{participationId}:
    parameters: [{$ref: '#/components/parameters/EventId'}, {$ref: '#/components/parameters/ParticipationId'}, {$ref: '#/components/parameters/IfMatch'}]
    patch:
      summary: Aprueba o rechaza una solicitud pendiente
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, required: [status], properties: {status: {type: string, enum: [confirmed, rejected]}}}}}
      responses:
        '200': {description: Participación resuelta, headers: {ETag: {$ref: '#/components/headers/ETag'}}, content: {application/json: {schema: {$ref: '#/components/schemas/Participation'}}}}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
        '428': {$ref: '#/components/responses/PreconditionRequired'}
  /events/{eventId}/invitations:
    parameters: [{$ref: '#/components/parameters/EventId'}]
    post:
      summary: Crea una invitación privada
      requestBody:
        content: {application/json: {schema: {type: object, properties: {expiresAt: {type: string, format: date-time}, maxUses: {type: integer, minimum: 1}}}}}
      responses:
        '201': {description: Invitación; el código solo se devuelve al creador, content: {application/json: {schema: {$ref: '#/components/schemas/Invitation'}}}}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /events/{eventId}/invitations/{invitationId}:
    parameters: [{$ref: '#/components/parameters/EventId'}, {$ref: '#/components/parameters/InvitationId'}]
    delete:
      summary: Revoca una invitación privada
      responses: {'204': {description: Invitación revocada}, '404': {$ref: '#/components/responses/NotFound'}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  headers:
    ETag: {schema: {type: string}, description: Versión opaca del recurso}
  parameters:
    EventId: {name: eventId, in: path, required: true, schema: {$ref: '#/components/schemas/Uuid'}}
    ParticipationId: {name: participationId, in: path, required: true, schema: {$ref: '#/components/schemas/Uuid'}}
    InvitationId: {name: invitationId, in: path, required: true, schema: {$ref: '#/components/schemas/Uuid'}}
...
    Cursor: {name: cursor, in: query, schema: {type: string}}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 50, default: 20}}
    IfMatch: {name: If-Match, in: header, required: false, schema: {type: string}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 1, maxLength: 100}}
  responses:
...
    Forbidden: {description: Operación no permitida, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
    NotFound: {description: Recurso inexistente u oculto, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
    Conflict: {description: Regla de negocio incumplida, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
    ValidationError: {description: Solicitud inválida, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
    PreconditionFailed: {description: ETag desactualizado, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
    PreconditionRequired: {description: Falta If-Match, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
  schemas:
    Uuid: {type: string, format: uuid}
    GeoPoint:
      type: object
      required: [type, coordinates]
      properties: {type: {type: string, enum: [Point]}, coordinates: {type: array, minItems: 2, maxItems: 2, items: {type: number}}}
    PublicProfile: {type: object, required: [id, alias], properties: {id: {$ref: '#/components/schemas/Uuid'}, alias: {type: string}}}
...
    EventPatch: {type: object, minProperties: 1, properties: {title: {type: string, minLength: 3, maxLength: 120}, description: {type: string, maxLength: 4000}, notes: {type: string, maxLength: 4000}, category: {type: string, enum: [sportWellbeing, cultureLeisure, learning, communityVolunteering, pets, networking]}, startsAt: {type: string, format: date-time}, durationMinutes: {type: integer, minimum: 15, maximum: 1440}, exactLocation: {$ref: '#/components/schemas/GeoPoint'}, accessMode: {type: string, enum: [direct, approval, privateInvitation]}, capacity: {type: integer, minimum: 1, nullable: true}}}
...
    EventDiscovery: {type: object, required: [id, title, description, category, startsAt, durationMinutes, accessMode, confirmedCount, availability, approximateArea, creator], properties: {id: {$ref: '#/components/schemas/Uuid'}, title: {type: string}, description: {type: string}, category: {type: string}, startsAt: {type: string, format: date-time}, durationMinutes: {type: integer}, accessMode: {type: string}, capacity: {type: integer, nullable: true}, confirmedCount: {type: integer}, availability: {type: string, enum: [available, full]}, approximateArea: {type: string}, distanceMeters: {type: integer}, creator: {$ref: '#/components/schemas/PublicProfile'}}}
    EventDetail: {allOf: [{$ref: '#/components/schemas/EventDiscovery'}, {type: object, properties: {notes: {type: string}, updatedAt: {type: string, format: date-time}, myParticipation: {type: string, enum: [pending, confirmed, rejected, abandoned]}, exactLocation: {$ref: '#/components/schemas/GeoPoint'}, confirmedParticipants: {type: array, items: {$ref: '#/components/schemas/PublicProfile'}}}}]}
    Participation: {type: object, required: [id, eventId, status, requestedAt], properties: {id: {$ref: '#/components/schemas/Uuid'}, eventId: {$ref: '#/components/schemas/Uuid'}, status: {type: string, enum: [pending, confirmed, rejected, abandoned]}, requestedAt: {type: string, format: date-time}, resolvedAt: {type: string, format: date-time, nullable: true}}}
    Invitation: {type: object, required: [id, code], properties: {id: {$ref: '#/components/schemas/Uuid'}, code: {type: string}, expiresAt: {type: string, format: date-time, nullable: true}, maxUses: {type: integer, nullable: true}}}
...
    CursorPage: {type: object, required: [items, page], properties: {items: {type: array, items: {type: object}}, page: {type: object, required: [nextCursor], properties: {nextCursor: {type: string, nullable: true}}}}}
...
    ParticipantPage: {allOf: [{$ref: '#/components/schemas/CursorPage'}]}
    Problem: {type: object, required: [type, title, status, code], properties: {type: {type: string, format: uri}, title: {type: string}, status: {type: integer}, code: {type: string}, detail: {type: string}, fields: {type: object, additionalProperties: {type: string}}}}